chore: Linted src/lib dir and main.cpp

+++ lib/spear/Color.cpp
@@ -1,23 +1,30 @@
+/*
+ * Copyright (c) 2026 Maximilian Krebs
+ * All rights reserved.
+*/
+
+#include "Color.h"
 #include <iomanip>
 #include <cmath>
-#include "Color.h"
+#include <cstdio>
+#include <string>
 
-Color::Color(unsigned r, unsigned g, unsigned b){
+Color::Color(unsigned r, unsigned g, unsigned b) {
     this->red = r;
     this->green = g;
     this->blue = b;
 }
 
-Color Color::interpolate(Color a, Color b, double amount){
+Color Color::interpolate(Color a, Color b, double amount) {
     Color toCalculate = Color(0, 0, 0);
 
-    double ared = (double) a.red;
-    double agreen = (double) a.green;
-    double ablue = (double) a.blue;
+    double ared = a.red;
+    double agreen = a.green;
+    double ablue = a.blue;
 
-    double bred = (double) b.red;
-    double bgreen = (double) b.green;
-    double bblue = (double) b.blue;
+    double bred = b.red;
+    double bgreen = b.green;
+    double bblue = b.blue;
 
     double inpRed = ared + (bred - ared) * amount;
     double inpGreen = agreen + (bgreen - agreen) * amount;
@@ -30,11 +37,11 @@ Color Color::interpolate(Color a, Color b, double amount){
     return toCalculate;
 }
 
-unsigned long Color::toHexVal(Color c){
+uint64_t Color::toHexVal(Color c) {
     return ((c.red & 0xff) << 16) + ((c.green & 0xff) << 8) + (c.blue & 0xff);
 }
 
-std::string Color::toHtmlColor(Color c){
+std::string Color::toHtmlColor(Color c) {
     char hexColor[8];
     std::snprintf(hexColor, sizeof hexColor, "#%02x%02x%02x", c.red, c.green, c.blue);
